SWITCH TO {sfnetworks}
This is a major re-work of core functions to, now, rely on the {sfnetworks} package instead of {stplanr}. These are 'drop in' updates so existing code should work without changes. One key update is the reliance on {sfnetworks} `st_network_blend()` function to snap spatial points (start/end) to the nearest edge instead of the nearest existing node.